Flytter ut klasser til egen fil
